requirements.txt
allure-pytest
allure-python-commons
pytest
selenium
webdriver-manager